Fixed a bug and a potential compiler error
If you used this class in Processing, it would say “Cannot refer to the
non-final local variable defined in an enclosing scope”, referring to
radio in addRadioMenuItem(). I replaced radio with the just-added
element in radios and that solved it.

Fixed a bug where passing a new RadioMenuItem(“text”, true) to
addRadioMenuItem would set it “on” twice.